Complete overhaul of types. May not 100% function well, but at least they are converted

+++ include/JMath.h
@@ -1,27 +1,29 @@
 #ifndef JMATH_H
 #define JMATH_H
 
+#include "types.h"
+
 namespace JMath
 {
     // not a typo :(
     class TRandom_fast_
     {
         public:
-        TRandom_fast_(unsigned long);
+        TRandom_fast_(u32);
 
-        int get();
-        float get_ufloat_1();
+        u32 get();
+        f32 get_ufloat_1();
         
-        int seed; // _0
+        u32 seed; // _0
     };
 
     class TRandom_enough_
     {
         public:
-        void setSeed(unsigned long);
+        void setSeed(u32);
 
-        int _0;
-        int _4;
+        u32 _0;
+        u32 _4;
     };
 };
