+++ include/node/TGraphGroup.h
@@ -2,6 +2,7 @@
 #define TGRAPHGROUP_H
 
 #include "JDrama.h"
+#include "types.h"
 #include "node/TGraphWeb.h"
 
 using namespace JDrama;
@@ -11,13 +12,13 @@ class TGraphGroup
     public:
     TGraphGroup(void *);
 
-    void perform(unsigned long, TGraphics *);
+    void perform(u32, TGraphics *);
     TGraphWeb* getGraphByName(char const *graphName);
     void initGraphGroup();
 
     void* _0;
-    int _4;
-    int* _8;
+    u32 _4;
+    u32* _8;
     TGraphWeb* graphWeb; // _C
 };
