+++ include/spc/TSpcBinary.h
@@ -1,6 +1,8 @@
 #ifndef TSPCBINARY_H
 #define TSPCBINARY_H
 
+#include "types.h"
+
 class TSpcBinary
 {
     public:
@@ -9,13 +11,13 @@ class TSpcBinary
 
     void initUserBuiltin();
     void* getHeader() const;
-    void bindSystemDataToSymbol(char const *, unsigned long);
-    int* searchSymbol(char const *);
+    void bindSystemDataToSymbol(char const *, u32);
+    u32* searchSymbol(char const *);
     void init();
     
     
     void* spcBinary; // _0
-    int* vtable; // _4
+    u32* vtable; // _4
 };
