+++ include/spc/spc.h
@@ -9,14 +9,14 @@ using namespace JDrama;
 
 void SpcTrace(char const *, ...);
 
-void evStartMontemanFanfare(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evStartMontemanBGM(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evIsWaterMelonIsReached(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evInvalidatePad(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evOnNeutralMarioKey(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evWarpFrontToMario(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evAppear8RedCoinsAndTimer(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evAppearReadyGo(TSpcTypedInterp<TEventWatcher> *, unsigned long);
-void evSetEventForWaterMelon(TSpcTypedInterp<TEventWatcher> *, unsigned long);
+void evStartMontemanFanfare(TSpcTypedInterp<TEventWatcher> *, u32);
+void evStartMontemanBGM(TSpcTypedInterp<TEventWatcher> *, u32);
+void evIsWaterMelonIsReached(TSpcTypedInterp<TEventWatcher> *, u32);
+void evInvalidatePad(TSpcTypedInterp<TEventWatcher> *, u32);
+void evOnNeutralMarioKey(TSpcTypedInterp<TEventWatcher> *, u32);
+void evWarpFrontToMario(TSpcTypedInterp<TEventWatcher> *, u32);
+void evAppear8RedCoinsAndTimer(TSpcTypedInterp<TEventWatcher> *, u32);
+void evAppearReadyGo(TSpcTypedInterp<TEventWatcher> *, u32);
+void evSetEventForWaterMelon(TSpcTypedInterp<TEventWatcher> *, u32);
 
 #endif // SPC_H
